+++ src/components/pages/CartPage/Cart/Cart.tsx
@@ -1,15 +1,15 @@
 import { Link } from 'react-router-dom';
 
-import * as styled from './CartContainer.styled';
+import * as styled from './Cart.styled';
 
-import { useCartSizeValue } from '../../recoils/recoilCart';
+import { useCartSizeValue } from '../../../../recoils/recoilCart';
 
-import { CartActions } from './CartActions/CartActions';
-import { CartItemList } from './CartItemList/CartItemList';
+import { CartActions } from '../CartActions/CartActions';
+import { CartItemList } from '../CartItemList/CartItemList';
 
-import { PATH } from '../../constants';
+import { PATH } from '../../../../constants';
 
-export const CartContainer = () => {
+export const Cart = () => {
   const cartSize = useCartSizeValue();
 
   return (
