+++ src/components/pages/CartPage/CartPage.tsx
@@ -1,6 +1,6 @@
 import * as styled from './CartPage.styled';
 
-import { CartContainer } from '../../CartContainer/CartContainer';
+import { Cart } from './Cart/Cart';
 import { OrderSummary } from '../../OrderSummary/OrderSummary';
 import { PageTitle } from '../../styled/PageTitle';
 
@@ -9,7 +9,7 @@ export const CartPage = () => {
     <>
       <PageTitle>장바구니</PageTitle>
       <styled.Content>
-        <CartContainer />
+        <Cart />
         <OrderSummary />
       </styled.Content>
     </>
